+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 ### Changed
 
 - Minor refactorization to the main component, cleanup.
+- Makes `IonIconProp` generic, to allow custom icon names.
 
 ### Fixed